Rydder i OppdaterOppgaveService
